% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{plot.synthesis}
\alias{plot.synthesis}
\title{Plot Diagnostic for SYNTHESIS Model}
\usage{
\method{plot}{synthesis}(x, audit = NULL, ...)
}
\arguments{
\item{x}{An object of class "synthesis".}

\item{audit}{An optional "synthesis_audit" object to overlay TGEP guards.}

\item{...}{Additional arguments.}
}
\description{
Plot Diagnostic for SYNTHESIS Model
}
