% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{print.synthesis}
\alias{print.synthesis}
\title{Print Method for SYNTHESIS Model}
\usage{
\method{print}{synthesis}(x, ...)
}
\arguments{
\item{x}{An object of class "synthesis".}

\item{...}{Additional arguments.}
}
\description{
Print Method for SYNTHESIS Model
}
