% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{print.synthesis_stability}
\alias{print.synthesis_stability}
\title{Print Method for SYNTHESIS Stability}
\usage{
\method{print}{synthesis_stability}(x, ...)
}
\arguments{
\item{x}{An object of class "synthesis_stability".}

\item{...}{Additional arguments.}
}
\description{
Print Method for SYNTHESIS Stability
}
