% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\docType{package}
\name{SYNTHESIS-package}
\alias{SYNTHESIS}
\alias{SYNTHESIS-package}
\title{SYNTHESIS: Structured Yield-Neutral Trust-weighted Heterogeneity-Integrated Statistical Inference System}
\description{
A 7-Layer Meta-Analysis Methodology for Robust Evidence Synthesis.
}
\author{
Mahmood Ul Hassan
}
\keyword{internal}
