% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{synthesis_table}
\alias{synthesis_table}
\title{Generate Transparency Table for SYNTHESIS Model}
\usage{
synthesis_table(x)
}
\arguments{
\item{x}{An object of class "synthesis".}
}
\description{
Generate Transparency Table for SYNTHESIS Model
}
