% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{validate_synthesis_object}
\alias{validate_synthesis_object}
\title{Validate SYNTHESIS Object Schema}
\usage{
validate_synthesis_object(x)
}
\arguments{
\item{x}{Object to validate.}
}
\description{
Checks if an object conforms to the v1.0 specification.
}
